Created a `clone()` syscall wrapper.

+++ src/runtime/implementations/exit.cpp
@@ -2,7 +2,8 @@
 // vim: set ft=cpp:
 #include <runtime.h>
 
-extern "C" [[gnu::used]] void exit(isize exit_code) {
+// Linux requires an unsigned 8-byte integer exit code for this syscall.
+extern "C" [[gnu::used]] void exit(i8 exit_code) {
     // TODO: is `meta::decay_numeral()` needed here? Why?
     asm("syscall" : : "D"(meta::decay_numeral(exit_code)), "a"(60));
     __builtin_unreachable();  // This elides a ret instruction.
